ajustando la aplicacion para pasar todos los componentes a la ruta / y creando nueva tipografia ademas de crear archivo frontpage
