+++ src/App.jsx
@@ -1,12 +1,13 @@
 import NavBar from "./components/NavBar/NavBar"
-import Index from "./pages/Index/Index"
+import AppRoutes from './routes/AppRoutes'
+// import Index from "./pages/Index/Index"
 
 function App() {
 
   return (
     <>
       <NavBar />
-      <Index />
+      <AppRoutes />
 
     </>
   )
