+++ src/routes/AppRoutes.jsx
@@ -1,5 +1,6 @@
 import { Routes, Route } from 'react-router-dom'
 import PrivateRoute from './PrivateRoutes'
+import Index from '../pages/Index/Index'
 
 
 
@@ -8,7 +9,7 @@ const AppRoutes = () => {
 
     return (
         <Routes>
-            <Route path='/' />
+            <Route path='/' element={<Index />} />
             <Route element={<PrivateRoute />}>
             </Route>
             <Route path='*' element={<h1> NOTHING BUT YOU ARE INSIDE </h1>} />
