+++ src/styles/base/_typography.scss
@@ -1,5 +1,5 @@
 
-// url fuente google
+// url fuente google rowdie
 @import url('https://fonts.googleapis.com/css2?family=Rowdies:wght@300;400;700&display=swap');
 
 
@@ -20,6 +20,28 @@
   font-weight: 700;
   font-style: normal;
 }
+// url fuente google intel
+@import url('https://fonts.googleapis.com/css2?family=Inter:wght@100..900&display=swap');
+
+@mixin font-intel-light {
+    font-family: "Inter", sans-serif;
+  font-weight: 300;
+  font-style: normal;
+}
+
+@mixin font-intel-regular {
+  font-family: "Inter", sans-serif;
+  font-weight: 400;
+  font-style: normal;
+}
+
+@mixin font-intel-bold {
+  font-family: "Inter", sans-serif;
+  font-weight: 700;
+  font-style: normal;
+}
+
+
