+++ src/styles/styles.scss
@@ -12,5 +12,9 @@
 
 // PAGES
 
-// pagina portada
+// pagina index
 @import './pages/Index.scss';
+
+
+// pagina portada
+@import './pages/FrontPage.scss';
